1: Restored IIImplementEObjectAttributes tests
